feat: add general.el, dired, ediff, hercules
I added a ton of packages for this commit, I forgot to commit earlier.

General.el provides easy keymap definition, using prefix keymaps and
even self-defined keymaps, which is great, instead of using
make-sparse-keymap.

dired-collapse provides dired with the typical ide funcionality,
removing the directory step if it only has an unique file
inside. dired-collapse shows directory/file instead.

I added some ediff configuration, but I'm still not 100% there. It seems
that the process could be a ton more seemless, but instead I'm using the
vertical popup and it is not the best.

Hercules is the hydra.el killer, providing consistent which-key-like key
completions for any modes. Nevertheless, I find the package a bit
missleading, considering it is not really clear how and when the popup
will occur.