feat: add modus-themes and solarized configuration
I started to look for light themes because I intend to be
programming mostly in daylight, and it is recommended to have
a high contrast theme for doing so.

The options I finally came up with are Solarized (light-high-contrast)
and modus-themes-operandi (the light version), configured with
protesilaos options.

I'm trying out both, but Solarized resonates more with my style of
sordid gruvbox themes.