fix: remove custom variables from init.el
By default, emacs appends all the customize information to the bottom of
the init.el file, which is not very clean.

An easier solution is to add the (load-file custom-file) to the bottom
of init.el, passing the problem somewhere else. The file will be written
out automatically and should not be edited by hand.