feat: add qk/cycle-theme function
I added the qk/themes-list using a stackoverflow post that I found
really interesting for cycling themes:
https://emacs.stackexchange.com/questions/24088/make-a-function-to-toggle-themes

The idea was to add to the qk/themes-list all the themes that would be
tangled, in order to cycle through 2 themes (with optional support for
more).

Initially, the 0-indexed theme is loaded, and the qk/cycle-theme changes
to the next theme on the least, making the change to dark (when the time
of the day comes) pretty easy.