fix: set up sasl for my irc account
Add a new function to get the password from lastpass, as an effort to
start using lastpass as my password manager in emacs, considering I use
it everywhere else.

An important note is that VPNs are banned in some servers,
i.e. irc.libera.chat, so I recommend either using a bouncer, or just
being careful with what you do online, assume the conversation to be public.