feat: add Scroll-bar section
Add the Scroll-bar section, on which I removed the scroll-bar which is
enabled by default (this has been long in my config, just refiling) and
I also added the Smooth Scrolling section, which intends to fix the
problem of pixel scrolling inside Emacs using the package
'smooth-scroll'.

At the moment it has served me well, but I reserve myself the right to
change it in the future, I hope when Emacs:master merges GTK3 or X11
graphical support, improving the experience.