feat: change preview for consult-buffer and theme
Change the narrow-key configuration from :config to :custom in
use-package, no need to use setq.

Remove :preview-key for consult-theme, no need to preview the theme on
change, it half bugs Emacs. Also add a preview key to consult-buffer, I
find that previewing on every buffer, which loads the major mode. That
means that lsp-mode loads for the preview buffers.