feat: add emacs-everywhere
Add emacs-everywhere package, which opens a new frame and pastes the
input to the selected text area.

Also configure org-babel to remove the warnings on compilation of source
blocks. Change the config.org buffer's properties to remove the results
when I chose to compile the source blocks with C-c C-c.