feat: add real auto-save
Xah-lee provides a function to auto-save all files that have not yet
been saved when the frame has changed focus.

I find interesting that the focus-out-hook has been deprecated and now
we need to add to the end of a funcion, using the add-function
macro. Pretty weird implementation, I could really dig into it.