feat: add support for tree-sitter-mode
I already had the support for tree-sitter-mode connected to the csharp
mode, but I figured why not really try to configure it properly for all
the major modes that support it.

There is no harm in trying out a breaking change to sintax-highlighting
like tree-sitter, I find that the community has been receiving the
package happily.