feat: add mediator package
Mediator is the ideal middle person in the conversation between
xdg-mime-type applications and Emacs, on which I coudn't really find a
good way to open the files with their desktop counterparts. Another
option I tried was openning the file with a 'async-shell-command' which
isn't really a good solution.

Add change on lpass_login for it to work properly. It isn't really
documented properly. I logged the result of the login command to /var/log/syslog.