feat: add transient help-command
Add a transient for the cluttered help-command (“C-h”). This way, there
is a clear understanding of the commands that are available, and you are
not searching for oscure commands that you don’t use. Using transient,
you learn command quicker.

Untangle good-scroll, at the moment the pixel-scroll functionality is
not that good in Emacs, considering the lack of good graphic internals.