fix: remove treemacs load time warnings
Bind the treemacs-no-load-time-warnings variable, which for some reason
wasn’t even created. The maintainer has a condition to check whether the
variable is bound, rather than checking if it set to true or nil, pretty
weird.

Definining (using defvar) in the :init portion of the treemacs config
solved the problem, no more warnings on load-time.