feat: add live share in Emacs: crdt.el
crdt.el is a real-time collaborative editing environment for Emacs
using Conflict-free Replicated Data Types. A shared session is a place
that can contains multiple buffers (or files), and multiple users can
join to collaboratively edit those buffers (or files). Think about a
meeting room with some people working together on some papers.

Live share is one of the key features of the lastest VSCode build, and
this is one of the first options to achieve the same using Emacs.