feat: add start of `qtile` configuration
Add thte config.py file for qtile configuration, which is a tiling
window manager written and configured in Python. I plan on using this,
but I still need to figure out the driver situation.

The Nvidia Drivers don’t seem to work well with window managers that are
not default. An option would be to use the nouveau open-source drivers,
and try to fix the problems that way.