feat: update themes, add after-load-theme-hook
Work on making the configuration more OS independent, including encoding
system, expand-region, warnings and more.

Remove "Setting faces eaily", as we now use :custom-face from
use-package for that. Try is also covered by straight, using
straight-use-package as an interactive command.

Add visiting-buffer.el, which provides an easy way to rename and change
the buffer name when renaming, as it is always the case. Remove the
--group-directories-first option for dired, which is only for GNU ls.

Add a help hydra, which contains all the help commands using helpful.

Remove toolbar, scroll-bar and smooth scrolling, as they are now defined
in the early-init.el file, which loads the X11 resources before loading
the early elisp.

Add perspective.el, in order to provide different buffer lists and workspaces.